IGE: Refactor validation classes for ISO import (#1546)
