+++ ingrid-mdek-job/src/main/java/de/ingrid/mdek/job/mapping/validation/iso/ISO_19115_2003_SchemaValidator.java
@@ -20,37 +20,32 @@
  * limitations under the Licence.
  * **************************************************#
  */
-package de.ingrid.mdek.job.mapping.bawdmqs;
+package de.ingrid.mdek.job.mapping.validation.iso;
 
-import de.ingrid.mdek.MdekError;
 import de.ingrid.mdek.job.MdekException;
 import de.ingrid.mdek.job.mapping.ImportDataMapper;
+import de.ingrid.mdek.job.mapping.validation.iso.util.IsoImportValidationUtil;
 import de.ingrid.mdek.job.protocol.ProtocolHandler;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
+import org.w3c.dom.Document;
 
-import java.util.List;
+import static de.ingrid.mdek.job.mapping.validation.iso.util.IsoImportValidationUtil.ISO_ELEMENTS_RESOURCE_BUNDLE;
+import static de.ingrid.mdek.job.mapping.validation.iso.util.IsoImportValidationUtil.ISO_MESSAGES_RESOURCE_BUNDLE;
 
-public class MapperQueue<S, T> implements ImportDataMapper<S, T> {
-    private static final Log LOG = LogFactory.getLog(MapperQueue.class);
+/**
+ * Validator for validating ISO (19115:2003/Corrigendum 1:2006(E)) XML-files
+ * against the ISO-19139 metadata schema.
+ *
+ * @author Vikram Notay
+ */
+public final class ISO_19115_2003_SchemaValidator implements ImportDataMapper<Document, Document> {
 
-    private List<ImportDataMapper<S, T>> mappers;
-
-    @Override
-    public void convert(S source, T target, ProtocolHandler protocolHandler) throws MdekException {
-        for(ImportDataMapper<S, T> m: mappers) {
-            try {
-                m.convert(source, target, protocolHandler);
-            } catch(Exception ex) {
-                String msg = "Could not convert from ISO-19115 to IGC format";
-                LOG.error(msg, ex);
-                throw new MdekException(new MdekError(MdekError.MdekErrorType.IMPORT_PROBLEM, msg));
-            }
-        }
+    public ISO_19115_2003_SchemaValidator() {
     }
 
-    public void setMappers(List<ImportDataMapper<S, T>> mappers) {
-        this.mappers = mappers;
+    @Override
+    public void convert(Document sourceIso, Document igcIgnored, ProtocolHandler ph) throws MdekException {
+        IsoImportValidationUtil validator = new IsoImportValidationUtil(sourceIso, ph, ISO_ELEMENTS_RESOURCE_BUNDLE, ISO_MESSAGES_RESOURCE_BUNDLE);
+        validator.validateXmlSchema(sourceIso);
     }
 
 }
